# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "",
    "version": "1.0.2",
    "title": "IP Snapshot",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "apiteam@swagger.io"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "localhost:3000",
  "basePath": "",
...
  "schemes": ["http"],
  "paths": {
...
    "/api/Performance": {
      "get": {
        "tags": ["Performance"],
        "summary": "List of all Performance Matrix",
        "description": "List of all Performance Matrix",
        "operationId": "getPerformance Matrix",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [],
        "responses": {
          "400": {
            "description": "Please FIll all fields correctly"
          },
          "200": {
            "description": "Success with data"
          }
        }
      },
      "post": {
        "tags": ["Performance"],
        "summary": " New Performance Matrix",
        "description": "Add a new performance matrix into db",
        "operationId": "addPerformance",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Performance that needs to be added to the store",
            "required": false,
            "schema": {
              "$ref": "#/definitions/Performance"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Please fill all fields correctly"
          },
          "200": {
            "description": "Success with data"
          }
        }
      }
    },
...
    "/api/SubPerformance": {
      "get": {
        "tags": ["Sub-Performance"],
        "summary": "List of all Sub-Performances",
        "description": "List of all sub Performance Matrix",
        "operationId": "getSubPerformance Matrix",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [],
        "responses": {
          "400": {
            "description": "Please FIll all fields correctly"
          },
          "200": {
            "description": "Success with data"
          }
        }
      },
      "post": {
        "tags": ["Sub-Performance"],
        "summary": " New Sub-Performance Matrix",
        "description": "Add a new sub-performance matrix into db",
        "operationId": "addPerformance",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Performance that needs to be added to the store",
            "required": false,
            "schema": {
              "$ref": "#/definitions/subPerformance"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Please fill all fields correctly"
          },
          "200": {
            "description": "Success with data"
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "petstore_auth": {
      "type": "oauth2",
      "authorizationUrl": "https://petstore.swagger.io/oauth/authorize",
      "flow": "implicit",
      "scopes": {
        "write:pets": "modify pets in your account",
        "read:pets": "read your pets"
      }
    },
    "api_key": {
      "type": "apiKey",
      "name": "api_key",
      "in": "header"
    }
  },
  "definitions": {
...
    "Performance": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "percentage": {
          "type": "integer"
        },
        "isActive": {
          "type": "integer"
        }
      }
    },
    "subPerformance": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "parentId": {
          "type": "integer"
        },
        "isActive": {
          "type": "integer"
        }
      }
    },
...
  },
...
}